clickbait_multimodal_refactor_v12.text_primary/requirements.txt
# Core
numpy>=1.24
pandas>=2.1
scikit-learn>=1.3
lightgbm>=4.0
# Deep Learning (optional heavy)
torch>=2.1
torchvision>=0.16
timm>=0.9.12
transformers>=4.44.0
open_clip_torch>=2.24.0
Pillow>=10.0
# Web demo
flask>=3.0
waitress>=3.0
# Config
pyyaml>=6.0.1

# Tuning & Progress
optuna>=3.6
tqdm>=4.66
